requirements.txt
opencv-python==4.10.0
Pillow==10.2.0
pyzbar==0.1.9
PyPDF2==3.0.1
pandas==2.2.2